requirements/test.txt
Django>=1.7
Pillow
django-crispy-forms